Testing changes into develop (#126)
* add integrity value decrement over time with timer options

* add an initial integrity decrement of 1

* implemented automatic integrity decrement
adjusted game parameters

* Edited: book_riddle table height
Edited: Taskspawn immediately starting

* add sounds for riddles when failing and succeeding

* add balancing for leaks riddle

* adding three different times for each difficulty to MedicalDisaster.cs

* moving setting of each time from constructor to initialize

* adds game exit task, which helps the user to navigate to the game exit location
- removes all active tasks, when the game is over

* changes game task integrity value to float

* balances integrity values of tasks button-sequence, asteroid-shooter, storage-riddle, energy-core to evaluated values
- removes timer from storage riddle

* fixes uninitialized difficulty value in game timer start function, by moving difficulty initialization to awake invocation

* disables door controller on game exit room

* changed: asteroid shooter controls to trigger right side

* prevents multiple metric export in the same game

* changed: added all factorys and changes to asteroid shooter in scene

* add medical disaster balancing

* removed debug message

* Fixed: sequence share of multiple riddles, while color of cube doesnt change

* translates mixing ingredients to german
- adds halo to plant to water
- changes color description of yellow to orange, to adapt to the look of the shader

* balancing mixing ingredients

* adds timer reference to game end procedure

* balancing marble gravity

* Added: book riddle description

* adjusts integrity values for the book riddle

* changes task priority of tasks

* removes duplicated phrase on game end info text

* adding rolling sound to sound-manager

* adding threshold to activate rolling sound

---------

+++ Assets/Scripts/Game/Tasks/TimerTask.cs
@@ -12,7 +12,7 @@ namespace Game.Tasks
             string taskName,
             string taskDescription,
             GameTaskType taskType,
-            int integrityValue = k_DefaultIntegrityValue
+            float integrityValue = k_DefaultIntegrityValue
         ) : base(taskName: taskName, taskDescription: taskDescription, taskType: taskType, integrityValue: integrityValue)
         {
             remainingTime = initialTimerTime;
